[Fix] #244 회원가입 DBTI 타이밍 오류 수정 및 사진 바텀시트 위치 개선
