+++ DOKI/Presentation/DBTI/View/DBTIResultView.swift
@@ -66,7 +66,7 @@ struct DBTIResultView: View {
         }
         .onAppear {
             if viewModel.dbtiEntryContext == .myPage {
-                viewModel.fetchDBTIResult(petId: 19)
+                viewModel.fetchDBTIResult()
             }
         }
     }
