+++ DOKI/Presentation/Register/View/RegisterView.swift
@@ -6,26 +6,30 @@
 //
 
 import SwiftUI
+import PhotosUI
 
 struct RegisterView: View {
     @ObservedObject var viewModel: RegisterViewModel
-    
+
     @EnvironmentObject var authManager: AuthManager
     @Environment(\.dismiss) private var dismiss
-    
+
     @Binding var hasCompletedRegister: Bool
     @Binding var showDBTIStart: Bool
-    
+
+    @State private var isShowPhotoSheet = false
+    @State private var selectedPhotoItem: PhotosPickerItem?
+
     var body: some View {
         VStack(spacing: 0) {
             currentStepView
-            
+
             Spacer()
-            
+
             mainButton
         }
         .overlay(alignment: .top) { progressBar }
-        .overlay(alignment: .top, content: {
+        .overlay(alignment: .top) {
             if viewModel.regionFlow == .map {
                 RegionMapView(
                     regionFlow: $viewModel.regionFlow,
@@ -35,7 +39,12 @@ struct RegisterView: View {
                     onResetSelection: { viewModel.resetRegionSelection() }
                 )
             }
-        })
+        }
+        .overlay {
+            if isShowPhotoSheet {
+                photoBottomSheet
+            }
+        }
         .ignoresSafeArea(.keyboard)
         .topNavigationView {
             BackButton {
@@ -57,6 +66,11 @@ struct RegisterView: View {
         .onTapGesture {
             UIApplication.shared.hideKeyboard()
         }
+        .onChange(of: selectedPhotoItem) { _, newItem in
+            guard let newItem else { return }
+            withAnimation { isShowPhotoSheet = false }
+            handleSelectedPhoto(newItem)
+        }
     }
 }
 
@@ -67,13 +81,13 @@ extension RegisterView {
             case .userProfile:
                 UserInfoView(viewModel: viewModel)
             case .petProfile:
-                PetInfoView(viewModel: viewModel)
+                PetInfoView(viewModel: viewModel, showPhotoSheet: $isShowPhotoSheet)
             case .region:
                 RegionView(viewModel: viewModel)
             }
         }
     }
-    
+
     private var progressBar: some View {
         ProgressView(value: 1.0/3 * Double(viewModel.currentStep.rawValue + 1))
             .frame(height: 2)
@@ -81,18 +95,84 @@ extension RegisterView {
             .tint(.defaultPrimary)
             .animation(.easeInOut, value: viewModel.currentStep)
     }
-    
+
     private var mainButton: some View {
         MainButton(text: viewModel.isLastStep ? "완료" : "다음", buttonState: viewModel.buttonDisabled ? .disabled : .active1) {
             if viewModel.isLastStep {
                 viewModel.registerUser()
-                hasCompletedRegister = true
-                showDBTIStart = true
             } else {
                 viewModel.goToNextStep()
             }
         }
         .padding(.horizontal, 16)
+        .onChange(of: viewModel.registerCompleted) { completed in
+            guard completed else { return }
+            hasCompletedRegister = true
+            showDBTIStart = true
+        }
+    }
+
+    private var photoBottomSheet: some View {
+        ZStack {
+            Color.black.opacity(0.3)
+                .ignoresSafeArea()
+                .onTapGesture {
+                    withAnimation { isShowPhotoSheet = false }
+                }
+
+            VStack {
+                Spacer()
+
+                VStack(spacing: 10) {
+                    VStack(spacing: 0) {
+                        PhotosPicker(selection: $selectedPhotoItem, matching: .images) {
+                            Text("갤러리")
+                                .mainDefault(color: .defaultPrimary)
+                                .frame(maxWidth: .infinity, minHeight: 56)
+                        }
+
+                        Divider()
+
+                        Button {
+                            withAnimation { isShowPhotoSheet = false }
+                            viewModel.petProfileImage = nil
+                        } label: {
+                            Text("기본 이미지")
+                                .mainDefault(color: .defaultPrimary)
+                                .frame(maxWidth: .infinity, minHeight: 56)
+                        }
+                    }
+                    .background(Color.white)
+                    .cornerRadius(16)
+
+                    Button {
+                        withAnimation { isShowPhotoSheet = false }
+                    } label: {
+                        Text("취소")
+                            .frame(maxWidth: .infinity, minHeight: 56)
+                            .background(Color.defaultPrimary)
+                            .foregroundStyle(.white)
+                            .cornerRadius(8)
+                    }
+                }
+                .padding(.horizontal, 16)
+                .transition(.move(edge: .bottom))
+            }
+        }
+        .animation(.easeInOut, value: isShowPhotoSheet)
+    }
+
+    private func handleSelectedPhoto(_ item: PhotosPickerItem) {
+        item.loadTransferable(type: Data.self) { result in
+            switch result {
+            case .success(let data):
+                guard let data, let image = UIImage(data: data) else { return }
+                DispatchQueue.main.async {
+                    viewModel.presignedURL(image)
+                }
+            default:
+                print("이미지를 불러오지 못했습니다.")
+            }
+        }
     }
 }
-
